# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sayq POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shift-audits:
    get:
      operationId: listShiftAudits
      tags: [shiftAudits]
      summary: List shift audits
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Shift audits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShiftAudit"
    post:
      operationId: createShiftAudit
      tags: [shiftAudits]
      summary: Close a shift and submit physical counts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftAuditInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftAudit"
  /shift-audits/expected:
    get:
      operationId: getExpectedStock
      tags: [shiftAudits]
      summary: Get current expected stock snapshot for closing a shift
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Expected stock snapshot
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /shift-audits/{id}:
    get:
      operationId: getShiftAudit
      tags: [shiftAudits]
      summary: Get a shift audit with reconciliation detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shift audit detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftAuditDetail"
  /shift-audits/{id}/verify:
    patch:
      operationId: verifyShiftAudit
      tags: [shiftAudits]
      summary: Validate audit and sync physical stock to inventory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftAudit"
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content streamed with correct Content-Type.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    InventoryItem:
      type: object
      required: [itemType, itemId, name, unit, currentStock]
      properties:
        itemType:
          type: string
          enum: [ingredient, semi_finished]
        itemId:
          type: integer
        name:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        minimalStock:
          type: ["number", "null"]
        costPricePerUnit:
          type: number
...
    ShiftAuditStockEntry:
      type: object
      required: [itemType, itemId, name, unit, quantity]
      properties:
        itemType:
          type: string
        itemId:
          type: integer
        name:
          type: string
        unit:
          type: string
        quantity:
          type: number
    ShiftAuditInput:
      type: object
      required: [branchId, actualStock]
      properties:
        branchId:
          type: integer
        cashierId:
          type: ["integer", "null"]
        shiftStart:
          type: ["string", "null"]
          format: date-time
        actualStock:
          type: array
          items:
            $ref: "#/components/schemas/ShiftAuditStockEntry"
        photoProofUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ShiftAudit:
      type: object
      required: [id, branchId, status, createdAt]
      properties:
        id:
          type: integer
        branchId:
          type: integer
        cashierId:
          type: ["integer", "null"]
        cashierName:
          type: ["string", "null"]
        shiftStart:
          type: ["string", "null"]
          format: date-time
        shiftEnd:
          type: ["string", "null"]
          format: date-time
        photoProofUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, verified, discrepancy]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        maxDiscrepancyPct:
          type: number
    ShiftAuditReconLine:
      type: object
      required: [itemType, itemId, name, unit, expected, actual, diff, diffPct, isWarning]
      properties:
        itemType:
          type: string
        itemId:
          type: integer
        name:
          type: string
        unit:
          type: string
        expected:
          type: number
        actual:
          type: number
        diff:
          type: number
        diffPct:
          type: number
        isWarning:
          type: boolean
    ShiftAuditDetail:
      type: object
      required: [id, branchId, status, createdAt, reconciliation]
      properties:
        id:
          type: integer
        branchId:
          type: integer
        cashierId:
          type: ["integer", "null"]
        cashierName:
          type: ["string", "null"]
        shiftStart:
          type: ["string", "null"]
          format: date-time
        shiftEnd:
          type: ["string", "null"]
          format: date-time
        photoProofUrl:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        maxDiscrepancyPct:
          type: number
        reconciliation:
          type: array
          items:
            $ref: "#/components/schemas/ShiftAuditReconLine"
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...