# lib/api-zod/src/generated/types/shiftAuditStockEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sayq POS API
 * OpenAPI spec version: 0.1.0
 */

export interface ShiftAuditStockEntry {
  itemType: string;
  itemId: number;
  name: string;
  unit: string;
  quantity: number;
}